+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### 🚀 New
+
+* Added `Trigger` class.
+
+
 ## 0.5.0 - December 21, 2024
 
 ### 💥 Breaking changes